requirements.txt
pydentic
pydantic[email]
python-dotenv

## AutoGen:
autogen-agentchat
autogen-core
autogen-ext
openai
tiktoken
autogen-ext[http-tool]

## Langchain and LangGraph:
langchain
langchain-openai
langchain-groq
langchain-core
langchain-community
pypdf
bs4
arxiv
pymupdf
wikipedia

langchain-community
faiss-cpu
langchain_google_genai
pypdf
langchain_huggingface
langchain_pinecone


## langgraph:
youtube_search
langchain-tavily
langgraph
langchain-chroma
chromadb
tweepy
duckduckgo-search
yfinance
googlemaps
langchain_experimental

